Fix type conversion by makeing timeout long (#164)
(cherry picked from commit 290002ad69ccb49e5daab7f83b877551fb8a3ca8)

+++ prbt_hardware_support/include/prbt_hardware_support/libmodbus_client.h
@@ -43,10 +43,10 @@ public:
   bool init(const char* ip, unsigned int port) override;
 
   //! @brief See base class.
-  void setResponseTimeoutInMs(unsigned int timeout_ms) override;
+  void setResponseTimeoutInMs(unsigned long timeout_ms) override;
 
   //! @brief See base class.
-  unsigned int getResponseTimeoutInMs() override;
+  unsigned long getResponseTimeoutInMs() override;
 
   //! @brief See base class.
   RegCont readHoldingRegister(int addr, int nb) override;
